Unable to customize UITabBarController via new UITabBarAppearance API
iOS 13 appears to introduce UITabBarAppearance UITabBarItemAppearance and UITabBarItemStateAppearance as a new mechanism for customizing the appearance of UITabBarController.

Unfortunately UITabBarAppearance (and related classes) only appear to customize the background of the tab bar, not the coloring of the tab bar items.

Supporting dark mode on iOS 13 must (for now) be accomplished using legacy APIs, a custom subclass of UITabBarController, and traitCollectionDidChange

Run on iOS 13 simulator to observe the lack of tab bar item customization.

Change

let tabBarStyling: TabBarStyling = .iOS13BrokenAPI
to

let tabBarStyling: TabBarStyling = .darkModeCompatibleWithOldAPI
in order to see dark mode compatibility via legacy API